==> 5 - process_inputs/2 - resico_application/6 - rule_training.ipynb <==
### Importing libraries

import numpy as np
import pandas as pd

from random import randint
from tabulate import tabulate
from IPython.display import clear_output

### Inspecting the data

DATA_FOLDER = "/home/kmilo/Dev/PhD/features-lab/data/so_data"
dataset = pd.read_csv("{}/Incomplete_SO_library_usage.csv".format(DATA_FOLDER))

dataset

classified_answers = dataset[dataset.Library_Used != '-']

classified_answers

unclassified_answers = dataset[dataset.Library_Used == '-']

unclassified_answers

unclassified_imports = unclassified_answers[unclassified_answers.Imports != "-"]
unclassified_not_imports = unclassified_answers[unclassified_answers.Imports == "-"]

unclassified_imports

unclassified_not_imports

### Analysing the unclassified dataset *with* import statements and matching them with the predicted library to already establish a classification

for index_row, row in unclassified_imports.iterrows():
    imports = list(set(row['Imports'].split(",")))
    predicted_libraries = row['Predicted_Library'].split(",")
    
    in_imports = False
    
    for predicted_import in predicted_libraries:
        for import_stmt in imports:
            if import_stmt.startswith(predicted_import):
                in_imports = True
    
    if in_imports:
        unclassified_imports.at[index_row, 'Library_Used'] = 'Y'
    else:
        unclassified_imports.at[index_row, 'Library_Used'] = 'N'

unclassified_imports

### Constructing the datasets from the imports processing to a posterior learning

# Querying the data and removing columns
data_imports_positive = unclassified_imports[unclassified_imports.Library_Used == 'Y'].drop(['AnswerID', 'Tags', 'Imports', 'Predicted_Library', 'Predicted_APIs', 'Total_Tokens'], axis=1)
data_imports_negative = unclassified_imports[unclassified_imports.Library_Used == 'N'].drop(['AnswerID', 'Tags', 'Imports', 'Predicted_Library', 'Predicted_APIs', 'Total_Tokens'], axis=1)

# Random sampling the negative to balance both datasets
data_imports_negative_sampled = data_imports_negative.sample(n=len(data_imports_positive), random_state=42)

# Concatenating the two dataframes
data_processed = pd.concat([data_imports_positive, data_imports_negative_sampled])
data_processed = data_processed.sample(frac=1)

X = pd.DataFrame(data_processed, columns=["Classes", "Methods", "Ambiguity"]).to_numpy()
y = pd.Series(data_processed.Library_Used, dtype="category").cat.codes.values

data_processed

### Detecting issues with the dataset

from scipy.spatial.distance import cosine

def closest_records(dataset, record, record_index, n=5):
    closest_indexes = []
    scores = []

    for index_row, row in dataset.iterrows():

        if index_row != record_index:
            record_vector = [record.Classes, record.Methods, record.Ambiguity]
            row_vector = [row.Classes, row.Methods, row.Ambiguity]

            similarity = 1 - cosine(record_vector, row_vector)

            if len(closest_indexes) < n:
                closest_indexes.append(index_row)
                scores.append(similarity)
            elif len(closest_indexes) == n:
                comparison = list(filter(lambda score: score < similarity, scores))

                if len(comparison) > 0:
                    index_score = scores.index(comparison[0])
                    closest_indexes[index_score] = index_row
                    scores[index_score] = similarity
    
    return closest_indexes, scores

checked_indexes = []
NEIGHBOURS = 10

for index_row, row in failure_rules.iterrows():

    if not index_row in checked_indexes:
        print("Checked records: {} / {}".format(len(checked_indexes), len(failure_rules)))
        
        records = closest_records(failure_rules, row, index_row, NEIGHBOURS)
        indexes_records = records[0]
        data_records = failure_rules.loc[indexes_records]

        print("Classes: {} Methods: {} Ambiguity: {} Library_Used: {}".format(row.Classes, row.Methods, row.Ambiguity, row.Library_Used))
        input_change = input("Would you like to modify the class of the record? (y/n): ")
        checked_indexes.append(index_row)
        
        current_class = row.Library_Used
        if current_class == "Y":
            modified_class = "N"
        else:
            modified_class = "Y"

        if input_change == "y":
            data_processed.loc[index_row, 'Library_Used'] = modified_class

            print("Related records:\n {}".format(tabulate(data_records, headers="keys", tablefmt="psql")))
            input_change_all_neigh = input("Would you like to modify the class of ALL neighbours to be the same as the reference record? (y/n) ")

            if input_change_all_neigh == "y":
                data_processed.loc[indexes_records, 'Library_Used'] = modified_class
            else:
                input_change_some_neigh = input("Would you like to modify the class of SOME neighbours to be the same as the reference record? (y/n) ")

                if input_change_some_neigh == "y":
                    numbers_change = input("Which indexes?. Please number them separated by space")
                    records_change = [int(number) for number in numbers_change.split(" ")]

                    data_processed.loc[records_change, 'Library_Used'] = modified_class
        else:
            print("Related records:\n {}".format(tabulate(data_records, headers="keys", tablefmt="psql")))
            input_change_all_neigh = input("Avoid also changes to the neighbours? (y/n) ")
            
            if input_change_all_neigh == "n":
                input_decision = input("Change by bulk (1) or by index (2)? (1/2)")
                
                if input_decision == "1":
                    records_positive = data_records[data_records.Library_Used == "Y"]
                    records_negative = data_records[data_records.Library_Used == "N"]
                    
                    input_decision2 = input("Y ==> N? or N ==> Y? (1/2)")
                    
                    if input_decision2 == "1":
                        data_processed.loc[list(records_positive.index), 'Library_Used'] = "N"
                    elif input_decision2 == "2":
                        data_processed.loc[list(records_negative.index), 'Library_Used'] = "Y"
                    
                if input_decision == "2":
                    numbers_change = input("Which indexes?. Please number them separated by space")
                    records_change = [int(number) for number in numbers_change.split(" ")]

                    data_processed.loc[records_change, 'Library_Used'] = modified_class
        
        checked_indexes += indexes_records
        print()

data_processed.to_csv("/home/kmilo/data_processed.csv", index=False)

### Loading the processed data

data_loaded = pd.read_csv("/home/kmilo/data_processed.csv")
data_loaded

### Balancing the loaded data

data_loaded_positive = data_loaded[data_loaded.Library_Used == 'Y']
data_loaded_negative = data_loaded[data_loaded.Library_Used == 'N']

# Random sampling the negative to balance both datasets
data_loaded_negative_sampled = data_loaded_negative.sample(n=len(data_loaded_positive), random_state=42)

# Concatenating the two dataframes
data_loaded_processed = pd.concat([data_loaded_positive, data_loaded_negative_sampled])
data_loaded_processed = data_loaded_processed.sample(frac=1)

data_loaded_processed

data_loaded_processed.to_csv("/home/kmilo/data_processed_balanced.csv", index=False)

### Rules (based on the imbalanced dataset)

The rules perform better in the imbalanced scenario, probably because they capture better the characteristics of an instance being predicted as Y or N, even though the imbalance of the data. This behaviour has been reported by at least two papers:

- The Impact of Class Rebalancing Techniques on the Performance and Interpretation of Defect Prediction Models by Chakkrit Tantithamthavorn; Ahmed E. Hassan; Kenichi Matsumoto
- Dealing with imbalanced data for interpretable defect prediction by Yuxiang Gao, Yi Zhu, Yu Zhao

def j48(clas, meth, amb):
    if amb == 0:
        if meth == 0:
            return 'N'
        return 'Y'
    else:
        if clas <= 4:
            if meth == 0:
                return 'N'
            else:
                if amb <= 1:
                    if clas <= 2:
                        return 'N'
                    return 'Y'
                return 'N'
        else:
            if amb <= 2:
                return 'Y'
            else:
                if clas <= 8:
                    if amb <= 3:
                        if clas <= 6:
                            return 'N'
                        return 'Y'
                    return 'N'
                else:
                    if amb <= 7:
                        if amb <= 5:
                            return 'Y'
                        else:
                            if clas <= 13:
                                return 'N'
                            return 'Y'
                    else:
                        if clas <= 18:
                            return 'N'
                        else:
                            if meth <= 19:
                                return 'Y'
                            return 'N'

### Inspecting the predictions made by the rule model

hits = 0
fails = 0

for _, row in data_processed.iterrows():
    prediction = j48(row.Classes, row.Methods, row.Ambiguity)
    
    if prediction == row.Library_Used:
        hits += 1
    else:
        fails += 1

print("Hits: {} Failures: {}".format(hits, fails))
print("Hits accuracy: {}".format(round(hits / (hits + fails), 2)))

selected_index_failures = []

for index_row, row in data_processed.iterrows():
    prediction = j48(row.Classes, row.Methods, row.Ambiguity)
    real_prediction = row.Library_Used 
    
    if prediction != real_prediction and real_prediction == 'Y':
        selected_index_failures.append(index_row)

failure_rules = data_processed.loc[selected_index_failures]
print(len(failure_rules))

failure_rules

### Predicting the outcome for the unclassified dataset

for index, (index_row, row) in enumerate(unclassified_not_imports.iterrows()):
    if index % 10_000 == 0:
        print("Predicting row {}".format(index))

    prediction = j48(row.Classes, row.Methods, row.Ambiguity)
    
    if prediction == 'Y':
        unclassified_not_imports.loc[index_row, 'Library_Used'] = 'Y'
    else:
        unclassified_not_imports.loc[index_row, 'Library_Used'] = 'N'

newly_classified_answers = unclassified_not_imports
newly_classified_answers

data_library_used = newly_classified_answers[(newly_classified_answers.Library_Used == 'Y')]
data_library_used

### Randomly inspecting some elements

def show_predictions(library):
    predictions = data_library_used[data_library_used.Predicted_Library == library]
    print("Number of predictions for library {}: {} / {}".format(library, len(predictions), len(data_library_used)))

    size_sample = 10

    random_indexes = np.random.choice(range(len(predictions)), size=size_sample, replace=False)
    sample_library = predictions.iloc[random_indexes]

    for _, row in sample_library.iterrows():
        print("Link: https://stackoverflow.com/questions/{}".format(row.AnswerID))

#### Guava

libraries = [
    "com.google.common",
    "org.jfree",
    "org.jsoup",
    "org.quartz",
    "org.apache.pdfbox",
    "org.apache.poi",
    "org.apache.http",
]

show_predictions(libraries[0])

#### JFreeChart

show_predictions(libraries[1])

#### JSoup

show_predictions(libraries[2])

#### Quartz

show_predictions(libraries[3])

#### PDFBox

show_predictions(libraries[4])

#### POI

show_predictions(libraries[5])

#### Http

show_predictions(libraries[6])

### Gathering all classified data to export and re-run LiFUSO on it

# Tagged answers with the name of the libraries, untagged answers but with import statements including library APIs and
# the latest classified data by RESICO + rules + thresholds

gathered_data = pd.concat([classified_answers,
                           unclassified_imports[unclassified_imports.Library_Used == 'Y'],
                           data_library_used
                          ])

gathered_data

gathered_data.to_csv("{}/Processed_SO_data.csv".format(DATA_FOLDER))
